Added stuff
Added two new parameters to init_glfw
- winWidth and winHeight

As of right now they are not being used because I had no way to pass
data to the program to specify what I wanted for height/width.

For the moment, window size is a static 1000px wide and the
corresponding height to maintain aspect ratio.